requirements.txt
mediapipe
python-dotenv